Blood-bank-backend/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
authentication==1.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Django==4.1.11
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-environ==0.11.2
django-sms==0.6.0
djongo==1.3.6
dnspython==2.4.2
environ==1.0
Flask==3.0.0
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
passlib==1.7.4
pillow==10.2.0
pycryptodome==3.19.0
PyJWT==2.8.0
pymongo==3.12.3
pyotp==2.9.0
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
sqlparse==0.2.4
twilio==8.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.5
Werkzeug==3.0.0
WTForms==3.0.1
yarl==1.9.2
